backend/requirements.txt
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
boto3==1.16.51
botocore==1.19.51
certifi==2020.12.5
cffi==1.14.4
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
click==7.1.2
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
defusedxml==0.6.0
dj-database-url==0.5.0
dj-rest-auth==2.1.2
Django==3.1.5
django-allauth==0.44.0
django-appconf==1.0.4
django-apptemplates==1.5
django-cors-headers==3.6.0
django-countries==7.0
django-crispy-forms==1.10.0
django-extensions==3.1.0
django-imagekit==4.0.2
django-modeltranslation==0.16.1
django-money==1.2.2
django-papertrail==1.1.9
django-storages==1.11.1
django-timezone-field==4.1.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
docutils==0.15.2
h11==0.12.0
hiredis==1.1.0
httptools==0.1.1
huey==2.3.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
jmespath==0.10.0
msgpack==1.0.2
oauthlib==3.1.0
pilkit==2.0
Pillow==8.1.0
psycopg2-binary==2.8.6
py-moneyed==0.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==2.0.0
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.5
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.3.3
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.12.1
typing-extensions==3.7.4.3
urllib3==1.26.2
uvicorn==0.13.3
uvloop==0.14.0
websockets==8.1
Werkzeug==1.0.1
whitenoise==5.2.0
zope.interface==5.2.0
